requirements.txt
python-dotenv
pandas
web3